Move inventory stuff to its own module, patch requests properly and remove Coroutinemock
