+++ pyproject.toml
@@ -14,7 +14,6 @@ requests = "^2.31.0"
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.3"
 pytest-asyncio = "^0.21.1"
-asynctest = "^0.13.0"
 
 [build-system]
 requires = ["poetry-core"]
